// apps/web/src/lib/api/generated/main/model/event-overview-response-dto.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * Lootlog API
 * The Lootlog API documentation
 * OpenAPI spec version: 1.0
 */
import type { EventOverviewResponseDtoHeroNpcsItem } from './event-overview-response-dto-hero-npcs-item';
import type { EventOverviewResponseDtoSchema0 } from './event-overview-response-dto-schema0';
import type { EventOverviewResponseDtoScoringMode } from './event-overview-response-dto-scoring-mode';

export interface EventOverviewResponseDto {
  id: string;
  guildId: string;
  name: string;
  world: string;
  active: boolean;
  /**
     * @nullable
     * @pattern ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
     */
  startsAt: string | null;
  /**
     * @nullable
     * @pattern ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
     */
  endsAt: string | null;
  /** @pattern ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$ */
  createdAt: string;
  /** @pattern ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$ */
  updatedAt: string;
  /** @nullable */
  basePointsPerKill?: number | null;
  /** @nullable */
  assignmentTimeoutMinutes?: number | null;
  /** @nullable */
  participationConfirmationMinutes?: number | null;
  /** @nullable */
  mapAssignmentCap?: number | null;
  /** @nullable */
  rulebookMarkdown?: string | null;
  scoringMode: EventOverviewResponseDtoScoringMode;
  scoringRules: (EventOverviewResponseDtoSchema0 | null) | null;
  heroNpcs: EventOverviewResponseDtoHeroNpcsItem[];
}